Rename to integrationserver, make it easier to use
This PR renames 'fakeserver' to 'integrationserver' as that seems to be
a more descriptive title.

It also reworks how to make use of this during development and adds some
clarifying docs. Specifically, it adds a `make dev` target which brings
up the new integrationserver, webpack, and Cypress in headed mode, and
clean things up on exit.