Set Integration Server to always run concurrently
The initial implementation of this checked the database features to see
if it can handle concurrent connections. The Sqlite backend is set to
'False' which would cause the integration tests to hang. (e.g.: A
browser makes multiple concurrent requests to a server when loading a
page.)

I suspect that this setting is left over cruft from when Sqlite wouldn't
handle concurrent connections. Newer versions of Sqlite do support this,
with the caveat that writes will lock the database:
https://sqlite.org/faq.html#q5
Django does seem to take this into account, by making available a
timeout setting for sqlite backends:
https://docs.djangoproject.com/en/2.2/ref/databases/#database-is-locked-errors

The purpose of this server is to facilitate integration testing and
local development, so it's unlikely that it will ever see concurrent
usage. Therefore, this commit sets the runserver command to use
threading. Using threads solves a bug where a client making concurrent
requests would hang.