Pin react-map-gl to <6.x
React Map GL has a dependency on Mapbox GL. Since v2.x of Mapbox GL, the
package is no longer open source, and it enforces a license that's
incompatable with Locus Tempus.

Per the react-map-gl maintainer's comment:
https://github.com/visgl/react-map-gl/issues/1248
this commit downgrades react-map-gl to be less than 6 which will ensure
that LT doesn't install the non-free version of Mapbox.