Refactor DeckGL layer aggregation, fix event creation bug
This commit refactors DeckGL layer aggregation by having the components
re-initialize the layers on each render. Previously, the DeckGL layers
were being saved to state, which started to cause inconsistencies in how
they behaved, and which variables they would bind to. E.g. a value
passed in to the layer would be the value at the time of layer creation,
not the value at render-time.

This refactor then permitted the click handlers to bind to the correct
copy of the values.