Infer layer on event creation
Depending on the state of the applicaton in the Activity space, the user
may or may not be able to add an event to the current active layer. This
commit adds logic to the add event method to infer which layer an event
should be added to if the user is not permitted to add to the current
active layer.