Implement author response event view flow
This commit fully implements the flow an author takes when looking at
response events. When they click on a response event they'll see the
detail view of that event. This commit implements the step back,
returning the author to the active response. Prior they would return to
the pane of all responses.